requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.5
django-phonenumber-field==7.0.2
phonenumbers==8.13.4
sqlparse==0.4.3
